+++ front/src/pages/TabletView/TabletView.tsx
